+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [1.0.1] - 1 8 2024
- - Add Assembly.
+ - Edite the root namespace.
 
 ## [1.0.0] - 1 8 2024
  - Initial Version.